src/requirements.txt
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.14.0
urllib3==2.0.6
user-agent==0.1.10
selenium==4.15.2
func-timeout==4.3.5
prometheus-client==0.17.1
requests==2.31.0
certifi==2023.7.22
websockets==11.0.3
xvfbwrapper==0.2.9
beautifulsoup4
websockets
fastapi
uvicorn
faker